# shared/generated/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333/api"
    }
  ],
  "paths": {
...
    "/article/": {
      "post": {
        "tags": [
          "Article"
        ],
        "description": "Create article",
        "operationId": "createArticle",
        "requestBody": {
          "description": "Article object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Article"
                  },
                  {
                    "$ref": "#/components/schemas/ArticlePro"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Article"
                    },
                    {
                      "$ref": "#/components/schemas/ArticlePro"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination size",
            "schema": {
              "type": "number"
            }
          }
        ],
        "description": "Get all articles",
        "operationId": "getArticles",
        "responses": {
          "200": {
            "description": "Articles response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Article"
                      },
                      {
                        "$ref": "#/components/schemas/ArticlePro"
                      }
                    ]
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/article/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "The ID of the article",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }
      ],
      "get": {
        "tags": [
          "Article"
        ],
        "description": "Get article by id",
        "operationId": "getArticleById",
        "responses": {
          "200": {
            "description": "Article response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Article"
                    },
                    {
                      "$ref": "#/components/schemas/ArticlePro"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Article"
        ],
        "description": "Update article by id",
        "operationId": "updateArticleById",
        "requestBody": {
          "description": "Article object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Article"
                  },
                  {
                    "$ref": "#/components/schemas/ArticlePro"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Article"
                    },
                    {
                      "$ref": "#/components/schemas/ArticlePro"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Article"
        ],
        "description": "Delete article by id",
        "operationId": "deleteArticleById",
        "responses": {
          "204": {
            "description": "Article deleted"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "hidden": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "themeId": {
            "type": "number"
          },
          "recommendedArticleIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "ArticleRecommandation": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "Recommandation": {
                  "type": "object",
                  "properties": {
                    "Article": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "themeId": {
                          "type": "number"
                        },
                        "bannerUrl": {
                          "type": "string"
                        },
                        "headerUrl": {
                          "type": "string"
                        },
                        "Theme": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "iconUrl": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "Theme": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "iconUrl": {
                "type": "string"
              }
            }
          },
          "bannerUrl": {
            "type": "string"
          },
          "headerUrl": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "hidden",
          "themeId"
        ]
      },
...
      "ArticlePro": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "salaireMoy": {
            "type": "string"
          },
          "tarificationMoy": {
            "type": "string"
          },
          "outils": {
            "type": "string"
          },
          "creation": {
            "type": "string"
          },
          "utilisateurs": {
            "type": "string"
          },
          "necessiteAudicance": {
            "type": "string"
          },
          "financement": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "hidden": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "themeId": {
            "type": "number"
          },
          "recommendedArticleIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "ArticleRecommandation": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "Recommandation": {
                  "type": "object",
                  "properties": {
                    "Article": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "title": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "themeId": {
                          "type": "number"
                        },
                        "bannerUrl": {
                          "type": "string"
                        },
                        "headerUrl": {
                          "type": "string"
                        },
                        "Theme": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number"
                            },
                            "name": {
                              "type": "string"
                            },
                            "iconUrl": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "Theme": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "name": {
                "type": "string"
              },
              "iconUrl": {
                "type": "string"
              }
            }
          },
          "banner": {
            "type": "string",
            "format": "binary"
          },
          "header": {
            "type": "string",
            "format": "binary"
          },
          "bannerUrl": {
            "type": "string"
          },
          "headerUrl": {
            "type": "string"
          }
        }
      },
...
    }
  }
}